requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.6.0
urllib3==2.2.3
wheel==0.45.1
